BioTranslator/pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "BioTranslator"
authors = [
    {name = "Hanwen Xu"},
    {name = "Yunwei Zhao"},
    {name = "Addie Woicik"},
    {name = "Hoifung Poon"},
    {name = "Russ B.Altman"},
    {name = "Sheng Wang"}
]
maintainers = [
    {name = "Hanwen Xu", email = "xuhw@cs.washington.edm"},
    {name = "Yunwei Zhao", email = "yzhao4@cs.washington.edu"},
]
readme = {file = "README.md", content-type="text/markdown"}
dynamic = ["version", "description"]
classifiers = [
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "Natural Language :: English",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.7",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
]
dependencies = [
    "sentence-transformers",
    "anndata>=0.6.4",
    "fbpca>=1.0",
    "umap-learn>=0.3.10",
    "matplotlib>=2.0.2",
    "numpy>=1.16.4",
    "scipy>=1.3.1",
    "scikit-learn>=0.21.3",
    "scanorama>=1.7.1",
    "torch==1.7.1",
    "pandas",
    "tqdm",
    "networkx",
    "genism",
    "nltk",
    "torchvision==0.8.2",
]
[project.urls]
Source = "https://github.com/ywzhao2002/BioTranslatorProject-master"

[project.scripts]